calendar/requirements.txt
bidict==0.22.1
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
websocket-client==1.5.1
